@@ -17,30 +17,38 @@
         }
     ],
     "support": {
-        "issues": "https://github.com/sebastianbergmann/php-code-coverage/issues"
+        "issues": "https://github.com/sebastianbergmann/php-code-coverage/issues",
+        "security": "https://github.com/sebastianbergmann/php-code-coverage/security/policy"
     },
     "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.2",
         "ext-dom": "*",
+        "ext-libxml": "*",
         "ext-xmlwriter": "*",
-        "phpunit/php-file-iterator": "^2.0.2",
-        "phpunit/php-token-stream": "^3.1.3 || ^4.0",
-        "phpunit/php-text-template": "^1.2.1",
-        "sebastian/code-unit-reverse-lookup": "^1.0.1",
-        "sebastian/environment": "^4.2.2",
-        "sebastian/version": "^2.0.1",
-        "theseer/tokenizer": "^1.1.3"
+        "nikic/php-parser": "^5.3.1",
+        "phpunit/php-file-iterator": "^5.1.0",
+        "phpunit/php-text-template": "^4.0.1",
+        "sebastian/code-unit-reverse-lookup": "^4.0.1",
+        "sebastian/complexity": "^4.0.1",
+        "sebastian/environment": "^7.2.0",
+        "sebastian/lines-of-code": "^3.0.1",
+        "sebastian/version": "^5.0.2",
+        "theseer/tokenizer": "^1.2.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.2.2"
+        "phpunit/phpunit": "^11.4.1"
     },
     "suggest": {
-        "ext-xdebug": "^2.7.2"
+        "ext-pcov": "PHP extension that provides line coverage",
+        "ext-xdebug": "PHP extension that provides line coverage as well as branch and path coverage"
     },
     "autoload": {
         "classmap": [
@@ -48,14 +56,13 @@
         ]
     },
     "autoload-dev": {
-        "files": [
-            "tests/TestCase.php",
-            "tests/_files/BankAccountTest.php"
+        "classmap": [
+            "tests/"
         ]
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "7.0-dev"
+            "dev-main": "11.0.x-dev"
         }
     }
 }
