@@ -16,13 +16,22 @@
         }
     ],
     "support": {
-        "issues": "https://github.com/sebastianbergmann/php-file-iterator/issues"
+        "issues": "https://github.com/sebastianbergmann/php-file-iterator/issues",
+        "security": "https://github.com/sebastianbergmann/php-file-iterator/security/policy"
     },
+    "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
+        "optimize-autoloader": true,
+        "sort-packages": true
+    },
+    "prefer-stable": true,
     "require": {
-        "php": ">=7.1"
+        "php": ">=8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^11.0"
     },
     "autoload": {
         "classmap": [
@@ -31,7 +40,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.0.x-dev"
+            "dev-main": "5.0-dev"
         }
     }
 }
