@@ -15,15 +15,31 @@
         }
     ],
     "support": {
-        "issues": "https://github.com/sebastianbergmann/php-text-template/issues"
+        "issues": "https://github.com/sebastianbergmann/php-text-template/issues",
+        "security": "https://github.com/sebastianbergmann/php-text-template/security/policy"
     },
+    "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
+        "optimize-autoloader": true,
+        "sort-packages": true
+    },
+    "prefer-stable": true,
     "require": {
-        "php": ">=5.3.3"
+        "php": ">=8.2"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^11.0"
     },
     "autoload": {
         "classmap": [
             "src/"
         ]
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-main": "4.0-dev"
+        }
     }
 }
-
