@@ -15,16 +15,20 @@
         }
     ],
     "support": {
-        "issues": "https://github.com/sebastianbergmann/php-timer/issues"
+        "issues": "https://github.com/sebastianbergmann/php-timer/issues",
+        "security": "https://github.com/sebastianbergmann/php-timer/security/policy"
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.1"
+        "php": ">=8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^11.0"
     },
     "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
         "optimize-autoloader": true,
         "sort-packages": true
     },
@@ -35,7 +39,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.1-dev"
+            "dev-main": "7.0-dev"
         }
     }
 }
