@@ -17,50 +17,46 @@
         }
     ],
     "support": {
-        "issues": "https://github.com/sebastianbergmann/phpunit/issues"
+        "issues": "https://github.com/sebastianbergmann/phpunit/issues",
+        "security": "https://github.com/sebastianbergmann/phpunit/security/policy"
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.2",
         "ext-dom": "*",
         "ext-json": "*",
         "ext-libxml": "*",
         "ext-mbstring": "*",
         "ext-xml": "*",
         "ext-xmlwriter": "*",
-        "doctrine/instantiator": "^1.3.1",
-        "myclabs/deep-copy": "^1.10.0",
-        "phar-io/manifest": "^2.0.3",
-        "phar-io/version": "^3.0.2",
-        "phpspec/prophecy": "^1.10.3",
-        "phpunit/php-code-coverage": "^7.0.12",
-        "phpunit/php-file-iterator": "^2.0.4",
-        "phpunit/php-text-template": "^1.2.1",
-        "phpunit/php-timer": "^2.1.2",
-        "sebastian/comparator": "^3.0.2",
-        "sebastian/diff": "^3.0.2",
-        "sebastian/environment": "^4.2.3",
-        "sebastian/exporter": "^3.1.2",
-        "sebastian/global-state": "^3.0.0",
-        "sebastian/object-enumerator": "^3.0.3",
-        "sebastian/resource-operations": "^2.0.1",
-        "sebastian/type": "^1.1.3",
-        "sebastian/version": "^2.0.1"
-    },
-    "require-dev": {
-        "ext-PDO": "*"
+        "myclabs/deep-copy": "^1.12.0",
+        "phar-io/manifest": "^2.0.4",
+        "phar-io/version": "^3.2.1",
+        "phpunit/php-code-coverage": "^11.0.7",
+        "phpunit/php-file-iterator": "^5.1.0",
+        "phpunit/php-invoker": "^5.0.1",
+        "phpunit/php-text-template": "^4.0.1",
+        "phpunit/php-timer": "^7.0.1",
+        "sebastian/cli-parser": "^3.0.2",
+        "sebastian/code-unit": "^3.0.1",
+        "sebastian/comparator": "^6.1.1",
+        "sebastian/diff": "^6.0.2",
+        "sebastian/environment": "^7.2.0",
+        "sebastian/exporter": "^6.1.3",
+        "sebastian/global-state": "^7.0.2",
+        "sebastian/object-enumerator": "^6.0.1",
+        "sebastian/type": "^5.1.0",
+        "sebastian/version": "^5.0.2"
     },
     "config": {
         "platform": {
-            "php": "7.2.0"
+            "php": "8.2.0"
         },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "suggest": {
-        "phpunit/php-invoker": "^2.0.0",
-        "ext-soap": "*",
-        "ext-xdebug": "*"
+        "ext-soap": "To be able to generate mocks based on WSDL files"
     },
     "bin": [
         "phpunit"
