@@ -9,11 +9,23 @@
             "email": "sebastian@phpunit.de"
         }
     ],
+    "support": {
+        "issues": "https://github.com/sebastianbergmann/code-unit-reverse-lookup/issues",
+        "security": "https://github.com/sebastianbergmann/code-unit-reverse-lookup/security/policy"
+    },
+    "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
+        "optimize-autoloader": true,
+        "sort-packages": true
+    },
+    "prefer-stable": true,
     "require": {
-        "php": ">=5.6"
+        "php": ">=8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^11.0"
     },
     "autoload": {
         "classmap": [
@@ -22,7 +34,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-main": "4.0-dev"
         }
     }
 }
