@@ -22,16 +22,25 @@
             "email": "bschussek@2bepublished.at"
         }
     ],
+    "support": {
+        "issues": "https://github.com/sebastianbergmann/comparator/issues",
+        "security": "https://github.com/sebastianbergmann/comparator/security/policy"
+    },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.1",
-        "sebastian/diff": "^3.0",
-        "sebastian/exporter": "^3.1"
+        "php": ">=8.2",
+        "sebastian/diff": "^6.0",
+        "sebastian/exporter": "^6.0",
+        "ext-dom": "*",
+        "ext-mbstring": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^11.4"
     },
     "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
         "optimize-autoloader": true,
         "sort-packages": true
     },
@@ -47,7 +56,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "3.0-dev"
+            "dev-main": "6.2-dev"
         }
     }
 }
