@@ -14,12 +14,24 @@
             "email": "mail@kore-nordmann.de"
         }
     ],
+    "support": {
+        "issues": "https://github.com/sebastianbergmann/diff/issues",
+        "security": "https://github.com/sebastianbergmann/diff/security/policy"
+    },
+    "prefer-stable": true,
+    "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
+        "optimize-autoloader": true,
+        "sort-packages": true
+    },
     "require": {
-        "php": ">=7.1"
+        "php": ">=8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5 || ^8.0",
-        "symfony/process": "^2 || ^3.3 || ^4"
+        "phpunit/phpunit": "^11.0",
+        "symfony/process": "^4.2 || ^5"
     },
     "autoload": {
         "classmap": [
@@ -33,7 +45,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "3.0-dev"
+            "dev-main": "6.0-dev"
         }
     }
 }
