@@ -2,7 +2,7 @@
     "name": "sebastian/environment",
     "description": "Provides functionality to handle HHVM/PHP environments",
     "keywords": ["environment","hhvm","xdebug"],
-    "homepage": "http://www.github.com/sebastianbergmann/environment",
+    "homepage": "https://github.com/sebastianbergmann/environment",
     "license": "BSD-3-Clause",
     "authors": [
         {
@@ -10,16 +10,23 @@
             "email": "sebastian@phpunit.de"
         }
     ],
+    "support": {
+        "issues": "https://github.com/sebastianbergmann/environment/issues",
+        "security": "https://github.com/sebastianbergmann/environment/security/policy"
+    },
     "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.1"
+        "php": ">=8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5"
+        "phpunit/phpunit": "^11.0"
     },
     "suggest": {
         "ext-posix": "*"
@@ -31,7 +38,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "4.2-dev"
+            "dev-main": "7.2-dev"
         }
     }
 }
