@@ -2,7 +2,7 @@
     "name": "sebastian/exporter",
     "description": "Provides the functionality to export PHP variables for visualization",
     "keywords": ["exporter","export"],
-    "homepage": "http://www.github.com/sebastianbergmann/exporter",
+    "homepage": "https://www.github.com/sebastianbergmann/exporter",
     "license": "BSD-3-Clause",
     "authors": [
         {
@@ -26,27 +26,39 @@
             "email": "bschussek@gmail.com"
         }
     ],
+    "support": {
+        "issues": "https://github.com/sebastianbergmann/exporter/issues",
+        "security": "https://github.com/sebastianbergmann/exporter/security/policy"
+    },
     "config": {
+        "platform": {
+            "php": "8.2.0"
+        },
         "optimize-autoloader": true,
         "sort-packages": true
     },
     "prefer-stable": true,
     "require": {
-        "php": ">=7.0",
-        "sebastian/recursion-context": "^3.0"
+        "php": ">=8.2",
+        "ext-mbstring": "*",
+        "sebastian/recursion-context": "^6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.0",
-        "ext-mbstring": "*"
+        "phpunit/phpunit": "^11.2"
     },
     "autoload": {
         "classmap": [
             "src/"
         ]
     },
+    "autoload-dev": {
+        "classmap": [
+            "tests/_fixture"
+        ]
+    },
     "extra": {
         "branch-alias": {
-            "dev-master": "3.1.x-dev"
+            "dev-main": "6.1-dev"
         }
     }
 }
